Health check

+++ scripts/controllers-initializer.js
@@ -2,8 +2,11 @@ const fs = require('fs');
 
 exports.init = () => {
   const path = './app/controllers/';
+  const remaining = ['healthCheck.js'];
   fs.readdirSync(path).forEach((elem, index) => {
-    fs.unlinkSync(`${path}/${elem}`);
+    if (!remaining.includes(elem)) {
+      fs.unlinkSync(`${path}/${elem}`);
+    }
   });
   fs.writeFileSync(`${path}/.keep`, '', 'utf8');
 };
